pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["jinwanseo <jinwanseo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.1"}
bcrypt = "^4.0.1"
psycopg2-binary = "^2.9.9"
pydantic-partial = "^0.5.2"
celery = "^5.3.5"
redis = "^5.0.1"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
